implement update

+++ utils/identifier.py
@@ -2,4 +2,4 @@ import datetime
 
 
 def ID():
-    return datetime.datetime.today().strftime('%Y%m%d%H%M%S')
+    return long(datetime.datetime.today().strftime('%Y%m%d%H%M%S'))
